+++ cli/download_source_test.go
@@ -10,11 +10,13 @@ import (
 	"strings"
 	"testing"
 
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+
+	"github.com/gruntwork-io/terragrunt/cli/tfsource"
 	"github.com/gruntwork-io/terragrunt/config"
 	"github.com/gruntwork-io/terragrunt/options"
 	"github.com/gruntwork-io/terragrunt/util"
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
 )
 
 func TestAlreadyHaveLatestCodeLocalFilePath(t *testing.T) {
@@ -223,56 +225,8 @@ func TestDownloadTerraformSourceFromLocalFolderWithManifest(t *testing.T) {
 
 }
 
-func TestSplitSourceUrl(t *testing.T) {
-	t.Parallel()
-
-	testCases := []struct {
-		name               string
-		sourceUrl          string
-		expectedRootRepo   string
-		expectedModulePath string
-	}{
-		{"root-path-only-no-double-slash", "/foo", "/foo", ""},
-		{"parent-path-one-child-no-double-slash", "/foo/bar", "/foo/bar", ""},
-		{"parent-path-multiple-children-no-double-slash", "/foo/bar/baz/blah", "/foo/bar/baz/blah", ""},
-		{"relative-path-no-children-no-double-slash", "../foo", "../foo", ""},
-		{"relative-path-one-child-no-double-slash", "../foo/bar", "../foo/bar", ""},
-		{"relative-path-multiple-children-no-double-slash", "../foo/bar/baz/blah", "../foo/bar/baz/blah", ""},
-		{"root-path-only-with-double-slash", "/foo//", "/foo", ""},
-		{"parent-path-one-child-with-double-slash", "/foo//bar", "/foo", "bar"},
-		{"parent-path-multiple-children-with-double-slash", "/foo/bar//baz/blah", "/foo/bar", "baz/blah"},
-		{"relative-path-no-children-with-double-slash", "..//foo", "..", "foo"},
-		{"relative-path-one-child-with-double-slash", "../foo//bar", "../foo", "bar"},
-		{"relative-path-multiple-children-with-double-slash", "../foo/bar//baz/blah", "../foo/bar", "baz/blah"},
-		{"parent-url-one-child-no-double-slash", "ssh://git@github.com/foo/modules.git/foo", "ssh://git@github.com/foo/modules.git/foo", ""},
-		{"parent-url-multiple-children-no-double-slash", "ssh://git@github.com/foo/modules.git/foo/bar/baz/blah", "ssh://git@github.com/foo/modules.git/foo/bar/baz/blah", ""},
-		{"parent-url-one-child-with-double-slash", "ssh://git@github.com/foo/modules.git//foo", "ssh://git@github.com/foo/modules.git", "foo"},
-		{"parent-url-multiple-children-with-double-slash", "ssh://git@github.com/foo/modules.git//foo/bar/baz/blah", "ssh://git@github.com/foo/modules.git", "foo/bar/baz/blah"},
-	}
-
-	for _, testCase := range testCases {
-		// Save a local copy in scope so all the tests don't run the final item in the loop
-		testCase := testCase
-		t.Run(testCase.name, func(t *testing.T) {
-			t.Parallel()
-
-			sourceUrl, err := url.Parse(testCase.sourceUrl)
-			require.NoError(t, err)
-
-			terragruntOptions, err := options.NewTerragruntOptionsForTest("testing")
-			require.NoError(t, err)
-
-			actualRootRepo, actualModulePath, err := splitSourceUrl(sourceUrl, terragruntOptions.Logger)
-			require.NoError(t, err)
-
-			assert.Equal(t, testCase.expectedRootRepo, actualRootRepo.String())
-			assert.Equal(t, testCase.expectedModulePath, actualModulePath)
-		})
-	}
-}
-
 func testDownloadTerraformSourceIfNecessary(t *testing.T, canonicalUrl string, downloadDir string, sourceUpdate bool, expectedFileContents string) {
-	terraformSource := &TerraformSource{
+	terraformSource := &tfsource.TerraformSource{
 		CanonicalSourceURL: parseUrl(t, canonicalUrl),
 		DownloadDir:        downloadDir,
 		WorkingDir:         downloadDir,
@@ -306,7 +260,7 @@ func testDownloadTerraformSourceIfNecessary(t *testing.T, canonicalUrl string, d
 }
 
 func testAlreadyHaveLatestCode(t *testing.T, canonicalUrl string, downloadDir string, expected bool) {
-	terraformSource := &TerraformSource{
+	terraformSource := &tfsource.TerraformSource{
 		CanonicalSourceURL: parseUrl(t, canonicalUrl),
 		DownloadDir:        downloadDir,
 		WorkingDir:         downloadDir,
