+++ cli/tfsource/types_test.go
@@ -1,0 +1,60 @@
+package tfsource
+
+import (
+	"net/url"
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+
+	"github.com/gruntwork-io/terragrunt/options"
+)
+
+func TestSplitSourceUrl(t *testing.T) {
+
+	t.Parallel()
+
+	testCases := []struct {
+		name               string
+		sourceUrl          string
+		expectedRootRepo   string
+		expectedModulePath string
+	}{
+		{"root-path-only-no-double-slash", "/foo", "/foo", ""},
+		{"parent-path-one-child-no-double-slash", "/foo/bar", "/foo/bar", ""},
+		{"parent-path-multiple-children-no-double-slash", "/foo/bar/baz/blah", "/foo/bar/baz/blah", ""},
+		{"relative-path-no-children-no-double-slash", "../foo", "../foo", ""},
+		{"relative-path-one-child-no-double-slash", "../foo/bar", "../foo/bar", ""},
+		{"relative-path-multiple-children-no-double-slash", "../foo/bar/baz/blah", "../foo/bar/baz/blah", ""},
+		{"root-path-only-with-double-slash", "/foo//", "/foo", ""},
+		{"parent-path-one-child-with-double-slash", "/foo//bar", "/foo", "bar"},
+		{"parent-path-multiple-children-with-double-slash", "/foo/bar//baz/blah", "/foo/bar", "baz/blah"},
+		{"relative-path-no-children-with-double-slash", "..//foo", "..", "foo"},
+		{"relative-path-one-child-with-double-slash", "../foo//bar", "../foo", "bar"},
+		{"relative-path-multiple-children-with-double-slash", "../foo/bar//baz/blah", "../foo/bar", "baz/blah"},
+		{"parent-url-one-child-no-double-slash", "ssh://git@github.com/foo/modules.git/foo", "ssh://git@github.com/foo/modules.git/foo", ""},
+		{"parent-url-multiple-children-no-double-slash", "ssh://git@github.com/foo/modules.git/foo/bar/baz/blah", "ssh://git@github.com/foo/modules.git/foo/bar/baz/blah", ""},
+		{"parent-url-one-child-with-double-slash", "ssh://git@github.com/foo/modules.git//foo", "ssh://git@github.com/foo/modules.git", "foo"},
+		{"parent-url-multiple-children-with-double-slash", "ssh://git@github.com/foo/modules.git//foo/bar/baz/blah", "ssh://git@github.com/foo/modules.git", "foo/bar/baz/blah"},
+	}
+
+	for _, testCase := range testCases {
+		// Save a local copy in scope so all the tests don't run the final item in the loop
+		testCase := testCase
+		t.Run(testCase.name, func(t *testing.T) {
+			t.Parallel()
+
+			sourceUrl, err := url.Parse(testCase.sourceUrl)
+			require.NoError(t, err)
+
+			terragruntOptions, err := options.NewTerragruntOptionsForTest("testing")
+			require.NoError(t, err)
+
+			actualRootRepo, actualModulePath, err := splitSourceUrl(sourceUrl, terragruntOptions.Logger)
+			require.NoError(t, err)
+
+			assert.Equal(t, testCase.expectedRootRepo, actualRootRepo.String())
+			assert.Equal(t, testCase.expectedModulePath, actualModulePath)
+		})
+	}
+}
